server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.3",
    "pyjwt>=2.11.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
